fix(mdx): satisfy clippy and 350-line CI gates
Allow disallowed alloc lints in MDX AST tests, and split MDX JSON/pretty
helpers into dedicated modules so the changed files stay within the
source-line limit.

#686
